First version of Human Detection in Python
This is the first version of human detection in Python. It uses Haar Cascade Classifier. The main code for implementing on server is in 'OpenCV_haarcascade.py' file. To run it you need to have 'haarcascade_fullbody.xml' file in the same location as this script. Also, there need to be installed 'OpenCV' library into Python-compiling enviroment.